requirement.txt
fastapi
uvicorn
openai
pinecone-client
pdfplumber
python-docx
tiktoken
python-dotenv
httpx
faiss-cpu